requirements.txt
monkeylearn~=3.6.0
Unidecode~=1.3.4
Markdown~=3.3.4
thefuzz~=0.19.0
pyfunctional~=1.4.3
google_api_python_client~=2.43.0
python-dotenv~=0.20.0
levenshtein~=0.18.1
requests~=2.27.1
google-search-results~=2.4.0
nltk~=3.7
vaderSentiment~=3.3.2
google-cloud-language~=2.4.1
Pebble~=4.6.3
glom~=22.1.0
selenium~=4.1.3